# schemas/openapi.yaml
...
openapi: 3.1.0
info:
  title: SauronID Core API
  version: 0.1.0
  description: |
    HTTP API of the SauronID Rust core (axum).

    Error envelope: 4xx responses from the central error type use
    `{"error":{"code":"snake_case","message":"...","fix":"..."}}` (being
    introduced in a parallel change). Some legacy handlers still return a
    plain-text error string body; both content types are documented on the
    shared error responses.

    Per-call signature (call-sig v2): operations marked with the
    `x-sauron-call-*` header parameters are behind the `require_call_signature`
    middleware. The client signs a canonical, length-prefixed payload under
    domain `sauron.call.v2` with the agent's registered Ed25519 PoP key. Signed
    fields, in order: version, agent_id, tenant_id, audience, method,
    target_uri, content_type, body_sha256, config_digest, timestamp_ms, nonce.
    Enforcement is gated by SAURON_REQUIRE_CALL_SIG (default on in production,
    advisory in development).

    Tenancy: the global tenancy layer extracts `x-sauron-tenant-id` (or the
    admin JWT `tnt` claim) and falls back to the `default` tenant. The header
    is documented on paths whose handlers consume the tenant id.

    Dev endpoints (`/dev/*`) are mounted only when
    SAURON_ENABLE_DEV_ENDPOINTS=1 and are marked `x-dev-only: true`.
servers:
  - url: http://localhost:3001
...
security: []
components:
  securitySchemes:
    adminKey:
      type: apiKey
      in: header
      name: x-admin-key
      description: >-
        Static admin key. Full-write keys allow any method; read-only keys
        allow GET/HEAD only.
    adminJwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        HS256 admin JWT. Scopes: admin:read (GET/HEAD), admin:write (mutating),
        admin:super / admin:full / * (super). Optional `tnt` claim restricts
        tenants.
    session:
      type: apiKey
      in: header
      name: x-sauron-session
      description: >-
        Tenant-bound HMAC user session issued by /user/auth or
        /user/auth/finish (format v2|tenant|key_image|exp|sig, 1h TTL).
  parameters:
    TenantId:
      name: x-sauron-tenant-id
      in: header
      required: false
      schema: { type: string, default: default }
      description: >-
        Tenant id consumed by the tenancy layer. Falls back to the `default`
        tenant for legacy callers. Admin JWTs may instead carry a `tnt` claim.
...
  schemas:
...
    StatsSubmitResponse:
      type: object
      additionalProperties: false
      required: [stored, latency_ms_verify, statement_hash]
      properties:
        stored: { type: boolean }
        latency_ms_verify: { type: integer, minimum: 0 }
        statement_hash: { type: string }
...
paths:
...
  /v1/stats/submit:
    post:
      tags: [v1-stats]
      summary: Submit a claimed metric value with a Groth16 integrity proof
      operationId: statsSubmit
      security: [{ adminKey: [] }, { adminJwt: [] }]
      parameters: [{ $ref: '#/components/parameters/TenantId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/StatsSubmission' }
      responses:
        '200':
          description: Stored after proof verification.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StatsSubmitResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /v1/stats/submit-transparent:
    post:
      tags: [v1-stats]
      summary: Submit a claimed metric value with a transparent (RISC-Zero) proof
      operationId: statsSubmitTransparent
      security: [{ adminKey: [] }, { adminJwt: [] }]
      parameters: [{ $ref: '#/components/parameters/TenantId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              description: TransparentStatsSubmission (TransparentProofPayload flattened).
              required: [metric_id, claimed_value, period_start, period_end, checkpoint_id, program_id, receipt_b64]
              properties:
                tenant_id: { type: string, default: "" }
                agent_id_or_none: { type: [string, "null"] }
                metric_id: { type: string, enum: [success_rate, error_rate, tool_call_count, cost_total] }
                claimed_value: { type: integer, format: int64 }
                period_start: { type: integer, format: int64 }
                period_end: { type: integer, format: int64 }
                checkpoint_id: { type: string }
                program_id: { type: string }
                receipt_b64: { type: string }
      responses:
        '200':
          description: Stored after proof verification.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StatsSubmitResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /v1/stats/cohort:
    get:
      tags: [v1-stats]
      summary: Query raw verified stats rows for a metric and period
      operationId: statsCohortQuery
      security: [{ adminKey: [] }, { adminJwt: [] }]
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - { name: metric_id, in: query, required: true, schema: { type: string } }
        - { name: period_start, in: query, required: true, schema: { type: integer, format: int64 } }
        - { name: period_end, in: query, required: true, schema: { type: integer, format: int64 } }
      responses:
        '200':
          description: Matching rows.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [rows, n]
                properties:
                  rows:
                    type: array
                    items:
                      type: object
                      additionalProperties: false
                      required: [tenant_id, metric_id, claimed_value, n_records, period_start, period_end, merkle_root, submitted_at]
                      properties:
                        tenant_id: { type: string }
                        agent_id_or_none: { type: [string, "null"] }
                        metric_id: { type: string }
                        claimed_value: { type: integer, format: int64 }
                        n_records: { type: integer, format: int64 }
                        period_start: { type: integer, format: int64 }
                        period_end: { type: integer, format: int64 }
                        merkle_root: { type: string }
                        submitted_at: { type: integer, format: int64 }
                  n: { type: integer, minimum: 0 }
        '401': { $ref: '#/components/responses/Unauthorized' }

  # ── /v1/cohort (admin-gated, DP-published surface) ──
  /v1/cohort:
    post:
      tags: [v1-cohort]
      summary: Create or update a cohort definition
      operationId: cohortUpsert
      security: [{ adminKey: [] }, { adminJwt: [] }]
      parameters: [{ $ref: '#/components/parameters/TenantId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CohortDefinition' }
      responses:
        '200':
          description: Stored definition.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CohortDefinition' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    get:
      tags: [v1-cohort]
      summary: List cohort definitions
      operationId: cohortList
      security: [{ adminKey: [] }, { adminJwt: [] }]
      parameters: [{ $ref: '#/components/parameters/TenantId' }]
      responses:
        '200':
          description: Definitions.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/CohortDefinition' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /v1/cohort/published:
    get:
      tags: [v1-cohort]
      summary: Fetch the DP-published cohort for a period
      description: Laplace noise per quartile under the cohort's epsilon budget; metrics below the k-anonymity threshold are suppressed.
      operationId: cohortPublished
      security: [{ adminKey: [] }, { adminJwt: [] }]
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - { name: cohort_id, in: query, required: true, schema: { type: string } }
        - { name: period_start, in: query, required: true, schema: { type: integer, format: int64 } }
        - { name: period_end, in: query, required: true, schema: { type: integer, format: int64 } }
      responses:
        '200':
          description: Published cohort.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PublishedCohort' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/cohort/{id}/budget:
    get:
      tags: [v1-cohort]
      summary: DP budget ledger for a cohort
      operationId: cohortBudgetGet
      security: [{ adminKey: [] }, { adminJwt: [] }]
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: Ledger entries.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: false
                  required: [cohort_id, metric_id, cycle_start, epsilon_spent, delta_spent, epsilon_cap, delta_cap, last_published]
                  properties:
                    cohort_id: { type: string }
                    metric_id: { type: string }
                    cycle_start: { type: integer, format: int64 }
                    epsilon_spent: { type: number }
                    delta_spent: { type: number }
                    epsilon_cap: { type: number }
                    delta_cap: { type: number }
                    last_published: { type: integer, format: int64 }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/cohort/{id}/budget/rotate:
    post:
      tags: [v1-cohort]
      summary: Rotate the DP budget cycle for a cohort
      operationId: cohortBudgetRotate
      security: [{ adminKey: [] }, { adminJwt: [] }]
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [new_cycle_start, new_epsilon_cap, new_delta_cap]
              properties:
                new_cycle_start: { type: integer, format: int64 }
                new_epsilon_cap: { type: number }
                new_delta_cap: { type: number }
                metric_ids:
                  oneOf:
                    - type: array
                      items: { type: string }
                    - type: "null"
      responses:
        '200':
          description: Rotation result.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [cohort_id, new_cycle_start, rotated]
                properties:
                  cohort_id: { type: string }
                  new_cycle_start: { type: integer, format: int64 }
                  rotated: { type: integer, minimum: 0 }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/cohort/{id}:
    parameters:
      - $ref: '#/components/parameters/TenantId'
      - { name: id, in: path, required: true, schema: { type: string } }
    get:
      tags: [v1-cohort]
      summary: Fetch a cohort definition
      operationId: cohortGet
      security: [{ adminKey: [] }, { adminJwt: [] }]
      responses:
        '200':
          description: Definition.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CohortDefinition' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [v1-cohort]
      summary: Delete a cohort definition
      operationId: cohortDelete
      security: [{ adminKey: [] }, { adminJwt: [] }]
      responses:
        '204': { description: Deleted. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ── /v1/admin/audit (admin-gated, tenant-scoped security events) ──
...